==> HMS unification/1. Clean_HMS.ipynb <==
import pandas as pd
import numpy as np
import re
import datetime
import os
pd.options.mode.chained_assignment = None  # default='warn'
import importlib
import global_settings
importlib.reload(global_settings)

setts = global_settings.Settings()

hms2 = pd.read_csv(setts.input_folder + '/MC_data.csv', dtype=str)
hms2.columns = setts.clean_columns(hms2)
hms2 = hms2.rename(
    columns = {#'jc#' : 'customer_code',
                'telephone_number' : 'mobile_phone'}
#               'Reordered Name' : 'customer_name'}
)
hms2['customer_name'] = hms2['reordered_name']
hms2 = hms2[['head_code', 'jc#', 'customer_code', 'customer_name', 'mobile_phone', 'customer_address', 'vin']]
hms2

# hms = pd.read_csv(setts.input_folder + '/List_of_MC_customers_doing_services_at_HEAD_01-09Dec2024.csv', dtype=str)
# hms.columns = setts.clean_columns(hms)
# hms = hms.rename(
#     columns = {#'jc#' : 'customer_code',
#                 'telephone_number' : 'mobile_phone',
#               'name' : 'customer_name'}
# )
# # hms['customer_name_old'] = hms['customer_name']
# hms = hms[['head_code', 'jc#', 'customer_code', 'customer_name', 'mobile_phone', 'customer_address', 'vin']]
# hms

# hms = pd.concat([hms, hms2], axis=0)
hms = hms2
hms

hms = hms.drop_duplicates(subset=['customer_code','customer_name','mobile_phone','customer_address','vin'])
hms

hms['mobile_phone'] = hms['mobile_phone'].apply(lambda x: setts.clean_phone(x))
hms = hms.dropna(subset=['mobile_phone'], how='all')
hms = hms.fillna('')

hms['customer_name'] = hms['customer_name'].apply(lambda x: setts.clean_name(x))
hms['customer_address'] = hms['customer_address'].apply(lambda x: setts.clean_address(x))

hms = hms[hms['mobile_phone'].apply(lambda x: len(str(x).strip()) > 0)]

hms = hms[['customer_code', 'customer_name', 'mobile_phone', 'customer_address', 'vin', 'head_code']]
hms

global_settings.replace_city(hms,'customer_address', 'city')
from unidecode import unidecode

hms['is_company'] = hms['customer_name'].fillna('').apply(lambda x: int(setts.mark_as_company(unidecode(x.lower()))))
hms['is_company'].value_counts()

hms

head_code_list = pd.read_excel(setts.input_folder + '/Danh_sach_dai_ly_20241216.xlsx', dtype=str, sheet_name='Sheet1')
global_settings.replace_city(head_code_list,'city', 'city')
head_code_list = head_code_list.set_index('head_code').to_dict()['city']
head_code_list

def extract_city_from_head(head, city, head_code_list):
    if len(city) == 0:
        if head in head_code_list:
            return head_code_list[head]
        else:
            return ''
    else:
        return city

hms['city'] = hms.apply(lambda x: extract_city_from_head(x['head_code'], x['city'], head_code_list), axis=1)
hms['customer_address'] = hms['customer_address'].fillna('') + ' ' + hms['city']
hms['customer_address'] = hms['customer_address'].str.strip()


# hms['customer_address'].apply(lambda x: setts.clean_address(x))
hms.to_csv(setts.tmp1_folder + '/cleaned_hms.csv', index=False)
hms

def count_blank_values(df, columns, column_combinations):
    blank_count = {}
    
    # Count blank values and calculate percentage for individual columns
    total_rows = len(df)
    for col in columns:
        blank_values = df[col].apply(lambda x: len(str(x).strip()) == 0).sum()
        blank_count[col] = {
            'count': blank_values,
            'percentage': (blank_values / total_rows) * 100 if total_rows > 0 else 0
        }
    
    # Count blank values and calculate percentage for specified combinations of columns
    for comb in column_combinations:
        combined_col_name = '&'.join(comb)
        blank_values = df[list(comb)].apply(lambda x: all(len(str(val).strip()) == 0 for val in x), axis=1).sum()
        blank_count[combined_col_name] = {
            'count': blank_values,
            'percentage': (blank_values / total_rows) * 100 if total_rows > 0 else 0
        }
        
    return blank_count

count_blank_values(hms, ['customer_name','mobile_phone','customer_address','vin'],
                      [['customer_name', 'customer_address']])

def count_non_blank_values(df, columns, column_combinations):
    non_blank_count = {}
    
    # Count non-blank values and calculate percentage for individual columns
    total_rows = len(df)
    for col in columns:
        non_blank_values = df[col].apply(lambda x: len(str(x).strip()) > 0).sum()
        non_blank_count[col] = {
            'count': non_blank_values,
            'percentage': (non_blank_values / total_rows) * 100 if total_rows > 0 else 0
        }
    
    # Count non-blank values and calculate percentage for specified combinations of columns
    for comb in column_combinations:
        combined_col_name = '&'.join(comb)
        non_blank_values = df[list(comb)].apply(lambda x: all(len(str(val).strip()) > 0 for val in x), axis=1).sum()
        non_blank_count[combined_col_name] = {
            'count': non_blank_values,
            'percentage': (non_blank_values / total_rows) * 100 if total_rows > 0 else 0
        }
        
    return non_blank_count

count_non_blank_values(hms, ['customer_name','mobile_phone','customer_address','vin'],
                      [['customer_name', 'customer_address']])

# import matplotlib.pyplot as plt

# def plot_pie_chart(data):
#     unique_values = {}
#     for item in data:
#         if item in unique_values:
#             unique_values[item] += 1
#         else:
#             unique_values[item] = 1

#     labels = unique_values.keys()
#     values = unique_values.values()

#     plt.figure(figsize=(8, 6))
#     plt.pie(values, labels=labels, autopct='%1.1f%%', startangle=140)
#     plt.axis('equal')  # Equal aspect ratio ensures that pie is drawn as a circle.

#     plt.title('Pie Chart of Values Count')
#     plt.show()

# # Example list of values
# # data_list = [1, 2, 2, 3, 3, 3, 4, 4, 4, 4]

# plot_pie_chart(hms['city'].to_list())
