==> HMS unification/1.2 Clean MH.ipynb <==
import pandas as pd
import numpy as np
import re
import datetime
import os
pd.options.mode.chained_assignment = None  # default='warn'
import importlib
import global_settings
from unidecode import unidecode
importlib.reload(global_settings)


setts = global_settings.Settings()

mh = pd.read_csv(setts.input_folder + '/MyHonda_20241220.csv', dtype=str)
mh.columns = setts.clean_columns(mh)
mh = mh.rename(
    columns = { 'customer_id' : 'customer_code',
                'name' : 'customer_name',
                'phone' : 'mobile_phone',
                'address' : 'customer_address'}
)
mh

# mh = mh.dropna(subset=['customer_name', 'mobile_phone', 'customer_address'], how='all')
mh = mh.drop_duplicates(subset=['customer_code','customer_name','mobile_phone','customer_address','vin'])
print(len(mh))

mh = mh.dropna(subset=['mobile_phone'], how='all')
mh['mobile_phone'] = mh['mobile_phone'].apply(lambda x: setts.clean_phone(x))

mh['province_id'] = mh['province_id'].astype(str).apply(lambda x: x.zfill(2))
mh['district_id'] = mh['district_id'].astype(str).apply(lambda x: x.zfill(3))
# Map district
district_map = pd.read_csv(setts.input_folder + '/district.csv', dtype=str)[['code', 'name']].rename(columns={'name': 'district_name', 'code' : 'district_id'})
mh = pd.merge(left=mh, right=district_map, how='left', on='district_id')

# Map province
province_map = pd.read_csv(setts.input_folder + '/province.csv', dtype=str)[['code', 'name']].rename(columns={'name': 'city', 'code' : 'province_id'})
mh = pd.merge(left=mh, right=province_map, how='left', on='province_id')

mh = mh.fillna('')

mh['customer_address'] = mh['customer_address'] + ' ' + mh['district_name'] + ' ' +  mh['city']
mh['customer_address'] = mh['customer_address'].str.strip()
mh['customer_name'] = mh['customer_name'].apply(lambda x: setts.clean_name(x))
mh['customer_address'] = mh['customer_address'].apply(lambda x: setts.clean_address(x))

mh = mh[mh['mobile_phone'].apply(lambda x: len(str(x).strip()) > 0)]


mh = mh[['customer_code', 'customer_name', 'mobile_phone', 'customer_address', 'vin', 'city']]
mh

mh['is_company'] = mh['customer_name'].fillna('').apply(lambda x: int(setts.mark_as_company(unidecode(x.lower()))))
mh['is_company'].value_counts()

mh[mh['is_company'] == 1] 

# global_settings.replace_city(mh,'city', 'city')
# mh

mh[:5000000].to_csv(setts.tmp1_folder + '/cleaned_mh_5M_1H.csv', index=False)
mh[5000000:].to_csv(setts.tmp1_folder + '/cleaned_mh_5M_2H.csv', index=False)


def count_blank_values(df, columns, column_combinations):
    blank_count = {}
    
    # Count blank values and calculate percentage for individual columns
    total_rows = len(df)
    for col in columns:
        blank_values = df[col].apply(lambda x: len(str(x).strip()) == 0).sum()
        blank_count[col] = {
            'count': blank_values,
            'percentage': (blank_values / total_rows) * 100 if total_rows > 0 else 0
        }
    
    # Count blank values and calculate percentage for specified combinations of columns
    for comb in column_combinations:
        combined_col_name = '&'.join(comb)
        blank_values = df[list(comb)].apply(lambda x: all(len(str(val).strip()) == 0 for val in x), axis=1).sum()
        blank_count[combined_col_name] = {
            'count': blank_values,
            'percentage': (blank_values / total_rows) * 100 if total_rows > 0 else 0
        }
        
    return blank_count

count_blank_values(mh, ['customer_name','mobile_phone','customer_address','vin'],
                      [['customer_name', 'customer_address']])

def count_non_blank_values(df, columns, column_combinations):
    non_blank_count = {}
    
    # Count non-blank values and calculate percentage for individual columns
    total_rows = len(df)
    for col in columns:
        non_blank_values = df[col].apply(lambda x: len(str(x).strip()) > 0).sum()
        non_blank_count[col] = {
            'count': non_blank_values,
            'percentage': (non_blank_values / total_rows) * 100 if total_rows > 0 else 0
        }
    
    # Count non-blank values and calculate percentage for specified combinations of columns
    for comb in column_combinations:
        combined_col_name = '&'.join(comb)
        non_blank_values = df[list(comb)].apply(lambda x: all(len(str(val).strip()) > 0 for val in x), axis=1).sum()
        non_blank_count[combined_col_name] = {
            'count': non_blank_values,
            'percentage': (non_blank_values / total_rows) * 100 if total_rows > 0 else 0
        }
        
    return non_blank_count

count_non_blank_values(mh, ['customer_name','mobile_phone','customer_address','vin'],
                      [['customer_name', 'customer_address']])
