==> HMS unification/2 Chia cluster OLD-NEW.ipynb <==
import pandas as pd
import numpy as np
import re
import datetime
import os
pd.options.mode.chained_assignment = None  # default='warn'
import importlib
import global_settings
importlib.reload(global_settings)
from unidecode import unidecode

setts = global_settings.Settings()

def clean_vin(input_string):
    extracted_values = re.findall(r'[0-9A-Z]', input_string)
    return ''.join(extracted_values)

# OLD DATA

# path = setts.output_folder + '_old'

############## REMEMBER TO CHANGE PATH OF OLD DATA #####################
path = '/data/workdirectory/03. HMS unification/1. 20250123/mc/output_old' 
############## REMEMBER TO CHANGE PATH OF OLD DATA #####################


df_deter = pd.read_csv(path + '/hms_mh_unified_match_vin_only.csv', dtype=str).drop(columns=['cleaned_address'])
df_case2 = pd.read_csv(path + '/hms_mh_unified_case2.csv', dtype=str).drop(columns=['cleaned_address'])
df_case3 = pd.read_csv(path + '/hms_mh_unified_case3.csv', dtype=str).drop(columns=['cleaned_address'])
df_case4 = pd.read_csv(path + '/hms_mh_unified_case4.csv', dtype=str).drop(columns=['cleaned_address'])


df_all_cases = pd.concat((df_deter, df_case2, df_case3, df_case4)).fillna('')
df_all_cases.columns = [i+'_old' for i in df_all_cases.columns.tolist()]
df_all_cases

# ONLY GET MH
df_all_cases = df_all_cases[df_all_cases['customer_code_old'].str.startswith('MH')]
df_all_cases

# NEW DATA

hms = pd.read_csv(setts.tmp2_folder + '/cleaned_hms_city.csv', dtype=str).fillna('')
hms['unification_id'] = ''
hms = hms[['unification_id', 'customer_code', 'vin', 'customer_name', 'customer_address', 'mobile_phone', 'district', 'city']]
hms['vin'] = hms['vin'].apply(lambda x: clean_vin(x.upper()))
col_backup = hms.columns.tolist()
hms['customer_code'] = 'HMS_' + hms['customer_code']
hms.columns = [i+'_new' for i in hms.columns.tolist()]

hms

# Match VIN only

match_vin_only = pd.merge(left=hms, right=df_all_cases, left_on='vin_new', right_on='vin_old', how='inner')
match_vin_only

match_vin_only_df_all_cases = match_vin_only[df_all_cases.columns]
match_vin_only_hms = match_vin_only[hms.columns]

match_vin_only_df_all_cases.columns = col_backup
match_vin_only_hms.columns = col_backup

print(len(match_vin_only_df_all_cases))
print(len(match_vin_only_hms))

match_vin_only_df_all_cases_same_code = df_all_cases[df_all_cases['customer_code_old'].isin(match_vin_only_df_all_cases['customer_code'])]
match_vin_only_df_all_cases_same_code.columns = col_backup
match_vin_only_df_all_cases = pd.concat([match_vin_only_df_all_cases, match_vin_only_df_all_cases_same_code], axis=0).drop_duplicates()

print(len(match_vin_only_df_all_cases_same_code))
print(len(match_vin_only_df_all_cases))

match_vin_only_hms

match_vin_only_df_all_cases

match_vin_only2 = pd.concat([match_vin_only_hms, match_vin_only_df_all_cases], axis=0)
match_vin_only2 = match_vin_only2.drop_duplicates().sort_values(by=['vin']).fillna('')
match_vin_only2['cluster'] = 0
match_vin_only2.to_csv(setts.tmp2_folder + '/hms_mh_match_vin_only.csv', index=False)

match_vin_only2

# Case 1 2 3 4

hms_proba = hms[~hms['customer_code_new'].isin(match_vin_only2['customer_code'])]
df_all_cases_proba = df_all_cases[~df_all_cases['customer_code_old'].isin(match_vin_only2['customer_code'])]

hms_proba.columns = col_backup
df_all_cases_proba.columns = col_backup
print(len(hms_proba))
print(len(df_all_cases_proba))

hms_df_all_cases_proba = pd.concat([hms_proba, df_all_cases_proba], ignore_index=True, axis=0)
hms_df_all_cases_proba['customer_name'] = hms_df_all_cases_proba['customer_name'].fillna('').astype(str).str.upper()
hms_df_all_cases_proba['customer_address'] = hms_df_all_cases_proba['customer_address'].fillna('').astype(str).str.upper()
hms_df_all_cases_proba

def check_null_case(name, address):
    # 5 Cases:
    # null name & address: 1
    # null name only: 2
    # null address only: 3
    # fulll name & address: 4
    # other: 0 (this should never happend)
    
    name = name.strip()
    address = address.strip()
    if len(name) == 0 and len(address) == 0:
        return 1
    elif len(name) == 0:
        return 2
    elif len(address) == 0:
        return 3
    elif len(name) > 0 and len(address) > 0:
        return 4
    else:
        return 0
    
hms_df_all_cases_proba['null_case'] = hms_df_all_cases_proba.fillna('').apply(lambda x: check_null_case(x['customer_name'], x['customer_address']), axis=1)
hms_df_all_cases_proba

print(hms_df_all_cases_proba['null_case'].value_counts())
print(hms_df_all_cases_proba['null_case'].value_counts() / len(hms_df_all_cases_proba) * 100)


case4_cluster = {'other' : 0, 'angiang' : 1, 'bacgiang' : 2, 'backan' : 3, 'baclieu' : 4, 'bacninh' : 5, 'bariavungtau' : 6, 'bentre' : 7, 'binhdinh' : 8, 'binhduong' : 9, 'binhphuoc' : 10, 'binhthuan' : 11, 'camau' : 12, 'cantho' : 13, 'caobang' : 14, 'daklak' : 15, 'daknong' : 16, 'danang' : 17, 'dienbien' : 18, 'dongnai' : 19, 'dongthap' : 20, 'gialai' : 21, 'hagiang' : 22, 'haiduong' : 23, 'haiphong' : 24, 'hanam' : 25, 'hanoi' : 26, 'hatinh' : 27, 'haugiang' : 28, 'hoabinh' : 29, 'hochiminh' : 30, 'hungyen' : 31, 'khanhhoa' : 32, 'kiengiang' : 33, 'kontum' : 34, 'laichau' : 35, 'lamdong' : 36, 'langson' : 37, 'laocai' : 38, 'longan' : 39, 'namdinh' : 40, 'nghean' : 41, 'ninhbinh' : 42, 'ninhthuan' : 43, 'phutho' : 44, 'phuyen' : 45, 'quangbinh' : 46, 'quangnam' : 47, 'quangngai' : 48, 'quangninh' : 49, 'quangtri' : 50, 'soctrang' : 51, 'sonla' : 52, 'tayninh' : 53, 'thaibinh' : 54, 'thainguyen' : 55, 'thanhhoa' : 56, 'thuathienhue' : 57, 'tiengiang' : 58, 'travinh' : 59, 'tuyenquang' : 60, 'vinhlong' : 61, 'vinhphuc' : 62, 'yenbai' : 63}
case4 = hms_df_all_cases_proba[hms_df_all_cases_proba['null_case'] == 4]
case4['cluster'] = case4['city'].apply(lambda x: case4_cluster[x])
case4

case4.to_csv(setts.tmp2_folder + '/hms_mh_case4.csv', index=False)

case2 = hms_df_all_cases_proba[hms_df_all_cases_proba['null_case'] == 2]
case2['cluster'] = 0
case2

case2.to_csv(setts.tmp2_folder + '/hms_mh_case2.csv', index=False)

import unicodedata
import re

def remove_accents(input_str):
    if not isinstance(input_str, str):
        return input_str
    input_str = input_str.replace('Đ', 'D').replace('đ', 'd')
    normalized_str = unicodedata.normalize('NFD', input_str)
    stripped_str = ''.join(
        char for char in normalized_str if unicodedata.category(char) != 'Mn'
    )
    return stripped_str

def get_letter_cluster(x):
    x = remove_accents(x)
    if x:
        if ' ' in x:
            first_character = x.strip().split(' ')[1][0].lower()
        else:
            first_character = x.strip()[0].lower()  
        if re.match(r'[a-z]', first_character):  
            return first_character
    return 26



def get_name_cluster(x, eng_lst = ['w', 'y', 'z', 'j', 'f'], vn_lst = ['t', 'v', 'h', 'n', 'd', 'l', 'p']):
    x = x.strip().lower()
    # Get first value of string
    n0 = remove_accents(x[0])
    
    # If first value isnt a vietnamese letter then return blank
    if n0 in eng_lst or not re.match(r'[a-v]', n0):
        return ''
    
    # If first letter is NOT a common letter in vietnamese alphabet then return it
    if n0 not in vn_lst:
        return n0
    
    # If first letter is a common letter in vietnamese alphabet then process it 
    # EXAMPLE:
    # NGUYEN VAN AN => NA
    # NGUYEN => N
    # NGUYEN あn => N
    # Check if x is a single word
    if ' ' not in x:
        return n0

    # If it's multi word
    x_s = x.split(' ')
    n1 = remove_accents(x_s[-1][0])

    # Check if second letter is an english letter or not
    if n1 in eng_lst or not re.match(r'[a-z]', n1):
        return n0
    else:
        return n0 + n1

def cluster_to_dict(cluster_lst):
    dct = {}
    for i in range(len(cluster_lst)):
        dct[cluster_lst[i]] = i
    return dct
    
case3 = hms_df_all_cases_proba[hms_df_all_cases_proba['null_case'] == 3]
case3['cluster'] = case3['customer_name'].apply(lambda x: get_name_cluster(x))
dct = cluster_to_dict(case3['cluster'].drop_duplicates().sort_values().tolist())
case3['cluster'] = case3['cluster'].apply(lambda x: dct[x] )

case3

# display(case3['cluster'].value_counts().reset_index().sort_values(by='index'))
case3.to_csv(setts.tmp2_folder + '/hms_mh_case3.csv', index=False)
