==> HMS unification/3. Cleansing & automate by cluster-match_vin_only.ipynb <==
import os
import pandas as pd
from unidecode import unidecode
import numpy as np
import re
import importlib
import global_settings
import hms_mh_cleansing_before_unification_function
import hms_mh_unification_after_cleansing_function


importlib.reload(global_settings)
importlib.reload(hms_mh_cleansing_before_unification_function)
importlib.reload(hms_mh_unification_after_cleansing_function)

setts = global_settings.Settings()
cleansing_function = hms_mh_cleansing_before_unification_function.cleansing_function
unification_old_new_function = hms_mh_unification_after_cleansing_function.unification_old_new_function

province = pd.read_csv(setts.input_folder + "/province.csv",dtype=str)
district = pd.read_csv(setts.input_folder + "/district.csv",dtype=str)
convert_phonenumber = pd.read_csv(setts.input_folder + "/convert_phonenumber.csv",dtype=str)

def create_folder_if_not_exists(folder_path):
    if not os.path.exists(folder_path):
        os.makedirs(folder_path)

# case = 'case4'
case = 'match_vin_only'
output_path = setts.output_folder + '/cluster_' + case
create_folder_if_not_exists(output_path)
print(output_path)

df_combi = pd.read_csv(setts.tmp2_folder + '/hms_mh_%s.csv' % case,dtype=str).fillna('')
df_combi

cluster_list = df_combi['cluster'].astype(int).drop_duplicates().sort_values().tolist()
print(cluster_list)

def generate_key(vin, name, address, phone, district, city, cut=False):
    if cut:
        name = hms_mh_cleansing_before_unification_function.clean_name_old_data(name)
        address = hms_mh_cleansing_before_unification_function.clean_address_old_data(address, district, city)
    
    return (vin+name+address+phone + district + city).lower().replace(' ', '')

df_combi['key'] = df_combi.apply(lambda x: generate_key(x['vin'], 
                                                          x['customer_name'], 
                                                          x['customer_address'], 
                                                          x['mobile_phone'], 
                                                          x['district'], 
                                                          x['city'], cut=True), axis=1)

df_old = df_combi[df_combi['unification_id'].str.len() != 0]
df_new = df_combi[df_combi['unification_id'].str.len() == 0]

print(len(df_old))
print(len(df_new))

# Go through each cluster
for cluster in cluster_list:
    print('=' * 50)
    print(f'Unifying case "{case}" cluster "{cluster}"')
    
    df_combi_cluster = df_combi[df_combi['cluster'] == str(cluster)]
    output_name = output_path + '/hms_mh_old_new_unified_cluster_%s.csv' % str(cluster)
    
    # Extract new and old data of cluster
    df_old_cluster = df_old[df_old['cluster'] == str(cluster)]
    df_new_cluster = df_new[df_new['cluster'] == str(cluster)]

    
    # Check if there are new data in cluster
    if len(df_new_cluster) == 0:
        print(f'No new data in cluster "{cluster}" of case "{case}", skipping')
        continue
        
    # Clean old & new data
    df_old_cluster, df_new_cluster = cleansing_function(df_old_cluster, 
                                                        df_new_cluster, 
                                                        province, district, 
                                                        convert_phonenumber)
    
    
    print('Number of rows in df_old: {:,}'.format(df_old_cluster.shape[0]))
    print('Number of rows in df_new: {:,}'.format(df_new_cluster.shape[0]))
    print('Expected number of rows after unification: {:,}'.format(df_new_cluster.shape[0] + df_old_cluster.shape[0]))
    
    # Unification
    df_unified_cluster = unification_old_new_function(df_old_cluster, 
                                                      df_new_cluster, 
                                                      convert_phonenumber)
    
    df_unified_cluster['key'] = df_unified_cluster.apply(lambda x: generate_key(x['vin_cleansing'], 
                                                                  x['name_cleansing'], 
                                                                  x['address_cleansing'], 
                                                                  x['phone_cleansing'], 
                                                                  x['district_group'], 
                                                                  x['city_group']), axis=1)
    
    df_unified_cluster = pd.merge(left = df_combi_cluster, right = df_unified_cluster, how='left', on='key')
    df_unified_cluster = df_unified_cluster[['unification_id_y', 'customer_code', 'vin', 'customer_name', 
                                           'customer_address', 'mobile_phone', 
                                           'name_cleansing', 'address_cleansing', 'phone_cleansing', 
                                           'district', 'city', 'cluster']]
    df_unified_cluster = df_unified_cluster.rename(columns = {'unification_id_y' : 'unification_id'})
    df_unified_cluster.to_csv(output_name, index=False)
    print('Actual number of rows after unification: {:,}'.format(df_unified_cluster.shape[0]))
    print('Output saved:', output_name)

df_unified_cluster

# df_unified_cluster_copy = df_unified_cluster.copy()
# df_unified_cluster_copy = df_unified_cluster_copy[['unification_id_y', 'customer_code', 'vin', 'customer_name', 
#                                            'customer_address', 'mobile_phone', 
#                                            'name_cleansing', 'address_cleansing', 'phone_cleansing', 
#                                            'district', 'city', 'cluster']]
# df_unified_cluster_copy = df_unified_cluster_copy.rename(columns = {'unification_id_y' : 'unification_id'})
# df_unified_cluster_copy.to_csv(output_name, index=False)

# df_unified_cluster_copy
