==> HMS unification/4. Generate unification id.ipynb <==
import pandas as pd
import numpy as np
import re
import datetime
import os
pd.options.mode.chained_assignment = None  # default='warn'
import importlib
import global_settings
importlib.reload(global_settings)


setts = global_settings.Settings()

case = 'case4'
# case = 'match_vin_only'
cluster_folder = setts.output_folder + '/cluster_' + case
# cluster_folder = setts.output_folder + '\deterministic_matching'
cluster_folder


glob_file = cluster_folder + '/hms_mh_old_new_unified_cluster_*.csv'
# glob_file = cluster_folder + '\\*.csv'
glob_file

import pandas as pd
import numpy as np
import re
from glob import glob
from unidecode import unidecode
import hashlib

def calculate_hash(input_str):
    sha256_hash = hashlib.sha256()
    sha256_hash.update(input_str.encode('utf-8'))
    hashed_bytes = sha256_hash.digest()

    hashed_str = ''.join(format(byte, '02x') for byte in hashed_bytes)
    return hashed_str

def filter_alphanumeric(input_string):
    return re.sub(r'[^a-zA-Z0-9]', '', input_string)

def get_uid(name, address, phone, vin):
    name = filter_alphanumeric(unidecode(name.lower()))
    address = filter_alphanumeric(unidecode(address.lower()))
    vin = filter_alphanumeric(unidecode(vin.lower()))
    
    st = ','.join([name,address,phone])
    return calculate_hash(st)

dfs = []
for file in glob(glob_file):
    print(file)
    df = pd.read_csv(file, dtype='str').fillna('')
    df = df[['unification_id', 'customer_code', 'vin', 'customer_name',
           'customer_address', 'mobile_phone', 'name_cleansing',
           'address_cleansing', 'phone_cleansing', 'district', 'city']]
    df['case'] = case
#     print(df.columns)
#     if len(df) > 0:
#         df.insert(0, 'unification_id', df.apply(lambda x: get_uid(x['customer_name'], x['customer_address'], x['mobile_phone'], x['vin']), axis=1))
#         uid_dict = df[['profile_id', 'unification_id']].drop_duplicates(subset=['profile_id']).set_index('profile_id').to_dict()['unification_id']
#         df['unification_id'] = df['profile_id'].apply(lambda x: uid_dict[x])
    dfs.append(df)
        
df_out = pd.concat(dfs, axis=0)
# df_out = df_out.drop(columns = ['profile_id'])
df_out

df_out.to_csv(setts.output_folder + f'/hms_mh_old_new_unified_{case}.csv', index=False)
