==> HMS unification/5. Hms_match (1).ipynb <==
import pandas as pd
import importlib
import global_settings
importlib.reload(global_settings)


setts = global_settings.Settings()

files = [
    setts.output_folder + '/hms_mh_old_new_unified_match_vin_only.csv',
    setts.output_folder + '/hms_mh_old_new_unified_case2.csv',
#     setts.output_folder + '/hms_mh_unified_case3.csv',
    setts.output_folder + '/hms_mh_old_new_unified_case4.csv',
]

dfs = []
for file in files:
    dfs.append(pd.read_csv(file, dtype=str))
    print(len(dfs[-1]))

unified_match_vin_only = pd.concat(dfs, axis=0)
mc_customer = pd.read_csv(setts.input_folder + '/MC_data.csv', dtype= str)

unified_match_vin_only.fillna('', inplace= True)
unified_match_vin_only

# Function to handle custom aggregation with unique values
def aggregate_group(df):
    names = ",".join(set(df.loc[df['customer_code'].str.startswith('MH', na=False), 'customer_name']))
    addresses = ",".join(set(df.loc[df['customer_code'].str.startswith('MH', na=False), 'customer_address']))
    mobile_phone = ",".join(set(df.loc[df['customer_code'].str.startswith('MH', na=False), 'mobile_phone']))
    vin = ",".join(set(df.loc[df['customer_code'].str.startswith('MH', na=False), 'vin']))
    id = ",".join(set(df.loc[df['customer_code'].str.startswith('MH', na=False), 'customer_code'].str.split('_').str[1]))
    return pd.Series([names, addresses, mobile_phone, vin , id])


result = unified_match_vin_only.groupby('unification_id').apply(aggregate_group).reset_index()
result.columns = ['unification_id', 'MH_customer_name', 'MH_customer_address', 'MH_phone_number','MH_vin', 'MH_id']

# Take all unification_id contains HMS code:
unification_id_n_customer_code = unified_match_vin_only[unified_match_vin_only['customer_code'].str.strip().str.contains('HMS')][['unification_id','customer_code']].copy()
unification_id_n_customer_code['customer_code'] = unification_id_n_customer_code['customer_code'].str.split('_').str[1]

df = result.merge(unification_id_n_customer_code, how = 'left', on = 'unification_id')
df.dropna(subset= 'customer_code', inplace= True)

final_df = mc_customer.merge(df[['MH_customer_name', 'MH_customer_address', 'MH_phone_number','MH_vin', 'MH_id','customer_code']], how  = 'left', left_on = 'Customer code', right_on= 'customer_code').fillna('')
final_df

final_df[['Customer code', 'Reordered Name', 'Telephone number',
       'Customer Address', 'Plate No', 'VIN',
       'MH_customer_name', 'MH_customer_address', 'MH_phone_number','MH_vin', 'MH_id']].to_excel(setts.output_folder + '/MC_data with_MH_data.xlsx', index=False)
